+++ pysat/tests/classes/cls_ci.py
